Added a timeout, made lookups a bit more efficient (#158)
Hash lookups that take longer than 10 seconds will now return a 503 status code with the `Retry-After` header set to 90 seconds.

This condition should only ever really be met when doing a lookup for a hash that has not yet been matched in the database.

Also; lookups that result in a zero-length string for the Publisher name will now no longer attempt to search for a Publisher and will skip this attribute, which should improve response times.